Account for other threshes in flag_bad_run.
